Super basic building "search" page
Fixed small formatting in docker stuff while troubleshooting. Change in
python server was a silent 400 response which is stupid and I need to
finish implementing that error case. Also logging...
Search page shows the results of /api/buildings. This is probably as
detailed as the search feature will get since search implementation
isn't the point of this project.